Split classes in files
