+++ tests/test_acpype.py
@@ -14,7 +14,10 @@ import os
 import shutil
 import pytest
 from pytest import approx
-from acpype_lib.acpype import ACTopol, MolTopol, _getoutput, init_main, version
+from acpype_lib import version
+from acpype_lib.acpype import init_main
+from acpype_lib.topol import ACTopol, MolTopol
+from acpype_lib.utils import _getoutput
 
 
 @pytest.mark.parametrize(
